requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
cffi==1.17.1
frozenlist==1.6.2
idna==3.10
msgspec==0.18.6
multidict==6.4.4
propcache==0.3.1
py-cord==2.6.1
pycares==4.9.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
yarl==1.20.0
